Consolidates hub list command
